+++ middleware.ts
@@ -1,0 +1,47 @@
+import { NextRequest, NextResponse } from "next/server";
+import { getSessionCookie } from "better-auth/cookies";
+ 
+export async function middleware(request: NextRequest) {
+	const { pathname } = request.nextUrl;
+	
+	// Skip middleware for excluded paths
+	if (
+		pathname === "/" ||
+		pathname.startsWith("/auth") ||
+		pathname.startsWith("/api") ||
+		pathname.startsWith("/_next") ||
+		pathname.startsWith("/public") ||
+		pathname === "/favicon.ico"
+	) {
+		return NextResponse.next();
+	}
+
+	const sessionCookie = getSessionCookie(request);
+ 
+    // THIS IS NOT SECURE!
+    // This is the recommended approach to optimistically redirect users
+    // We recommend handling auth checks in each page/route
+	if (!sessionCookie) {
+		const redirectUrl = new URL("/auth/signin", request.url);
+		redirectUrl.searchParams.set("redirectTo", request.nextUrl.pathname + request.nextUrl.search);
+		return NextResponse.redirect(redirectUrl);
+	}
+ 
+	return NextResponse.next();
+}
+ 
+export const config = {
+	matcher: [
+		/*
+		 * Match all request paths except for the ones starting with:
+		 * - / (landing page)
+		 * - /auth (authentication pages)
+		 * - /api (API routes)
+		 * - /_next/static (static files)
+		 * - /_next/image (image optimization files)
+		 * - /favicon.ico (favicon file)
+		 * - /public (public files)
+		 */
+		"/((?!$|auth|api|_next/static|_next/image|favicon.ico|public).*)",
+	],
+};
